Refactor to isolate Person responsabilities
