+++ app.rb
@@ -2,6 +2,7 @@ require_relative './student'
 require_relative './teacher'
 require_relative './book'
 require_relative './rental'
+require_relative './person_module.rb'
 
 class App
   def initialize(option)
@@ -10,6 +11,8 @@ class App
     @persons = []
     @rentals = []
   end
+  
+  include PersonModule
 
   # List all books.
   def list_books
@@ -25,44 +28,6 @@ class App
       puts "[#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
     @option.show_options
-  end
-
-  # Create a person (teacher or student, not a plain Person).
-  def create_person
-    puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    person = gets.chomp
-    case person
-    when '1'
-      create_student
-    when '2'
-      create_teacher
-    else
-      puts 'Write 1 or 2'
-    end
-    @option.show_options
-  end
-
-  # Create a student
-  def create_student
-    print 'Age: '
-    age = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-    print 'Has parent persmission? [Y/N]: '
-    parent_permission = gets.chomp.downcase == 'y'
-    @persons.push(Student.new(age, name, parent_permission))
-    puts 'Person created successfuly'
-  end
-
-  def create_teacher
-    print 'Age: '
-    age = gets.chomp
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    @persons.push(Teacher.new(age, name, specialization))
-    puts 'Person created successfully'
   end
 
   # Create a book.
@@ -106,9 +71,5 @@ class App
       puts "Date: #{rental.date}, Book \"#{rental.book.title}\" by #{rental.book.author}" if rental.person.id == id
     end
     @option.show_options
-  end
-
-  def exit
-    puts 'Thank you for using the app!'
   end
 end
