+++ main.rb
@@ -44,4 +44,8 @@ def main
   Options.new
 end
 
+def exit
+  puts 'Thank you for using the app!'
+end
+
 main
